refactor pv code for readability
